backend/pyproject.toml
[project]
name = "cohort-builder-backend"
version = "0.1.0"
description = "Cohort Builder Backend API"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "django>=5.0",
    "djangorestframework>=3.14",
    "django-cors-headers>=4.3",
    "psycopg2-binary>=2.9",
    "python-dotenv>=1.0",
    "celery>=5.3",
    "redis>=5.0",
    "django-celery-beat>=2.5",
    "django-celery-results>=2.5",
    "gunicorn>=21.2",
    "requests>=2.31",
    "cryptography>=41.0",
    "pandas>=2.0",
    "openai>=1.0",
    "numpy>=1.24",
    "google-cloud-storage>=2.10.0",
    "pydantic>=2.0.0",
    "scikit-learn>=1.3.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4",
    "pytest-django>=4.5",
    "black>=23.0",
    "ruff>=0.1",
    "ipython>=8.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["api", "cohort_backend"]

[tool.black]
line-length = 100
target-version = ['py310']

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE = "cohort_backend.settings"
python_files = ["test_*.py", "*_test.py"]
